refactor(menu): split message_scan.go into scan/run/config files
message_scan.go was 1,185 lines. Extract by concern (pure code movement):

- message_scan.go        - runGetScanType (the scan-type selection menu) (~251)
- message_scan_run.go    - runNewScanAll, determineStartMessage,
                           runUpdateNewscanPointers (scan execution + pointers)
- message_scan_config.go - runNewscanConfig (newscan area-config editor)

No behavior change. go vet clean, race-clean, 274 menu tests, 1428 total
across 49 packages.
